+++ src/ua.rs
@@ -13,6 +13,7 @@ mod client;
 mod client_config;
 mod continuation_point;
 mod data_type_definition;
+mod data_type_description;
 mod data_types;
 mod duration;
 mod event_id;
@@ -49,6 +50,7 @@ pub use self::{
     client::{Client, ClientState},
     continuation_point::ContinuationPoint,
     data_type_definition::DataTypeDefinition,
+    data_type_description::DataTypeDescription,
     data_types::*,
     duration::Duration,
     event_id::EventId,