+++ src/ua/data_types/node_id.rs
@@ -1,8 +1,8 @@
 use std::{ffi::CString, fmt, hash, str};
 
 use open62541_sys::{
-    UA_NODEID_NULL, UA_NODEID_NUMERIC, UA_NODEID_STRING_ALLOC, UA_NodeId_hash, UA_NodeId_parse,
-    UA_NodeId_print, UA_NodeIdType,
+    UA_NODEID_BYTESTRING_ALLOC, UA_NODEID_GUID, UA_NODEID_NULL, UA_NODEID_NUMERIC,
+    UA_NODEID_STRING_ALLOC, UA_NodeId_hash, UA_NodeId_parse, UA_NodeId_print, UA_NodeIdType,
 };
 
 use crate::{DataType, Error, ua};
@@ -60,6 +60,53 @@ impl NodeId {
         Self(inner)
     }
 
+    /// Creates GUID node ID.
+    #[must_use]
+    pub fn guid(ns_index: u16, guid: ua::Guid) -> Self {
+        let inner = unsafe { UA_NODEID_GUID(ns_index, guid.into_raw()) };
+        debug_assert_eq!(
+            inner.identifierType,
+            UA_NodeIdType::UA_NODEIDTYPE_GUID,
+            "new node ID should have GUID type"
+        );
+
+        Self(inner)
+    }
+
+    /// Creates byte string node ID.
+    ///
+    /// # Panics
+    ///
+    /// The byte string identifier must not contain any NUL bytes.
+    #[must_use]
+    pub fn byte_string(ns_index: u16, byte_string: &[u8]) -> Self {
+        // Unfortunately, `UA_NODEID_BYTESTRING_ALLOC` requires a NUL-terminated C `char` pointer to
+        // the buffer. This imposes the restriction that no node IDs with NUL bytes can be created.
+        let byte_string =
+            CString::new(byte_string).expect("node ID byte string does not contain NUL bytes");
+
+        // String allocation can fail but `UA_NODEID_BYTESTRING_ALLOC` does not tell us this when it
+        // happens. Instead, we end up with a well-defined node ID that has an empty byte string.
+        let inner = unsafe { UA_NODEID_BYTESTRING_ALLOC(ns_index, byte_string.as_ptr()) };
+        debug_assert_eq!(
+            inner.identifierType,
+            UA_NodeIdType::UA_NODEIDTYPE_BYTESTRING,
+            "new node ID should have byte string type"
+        );
+
+        // SAFETY: We have checked that we have this enum variant.
+        let identifier = unsafe { inner.identifier.byteString.as_ref() };
+        if !byte_string.is_empty() && (identifier.data.is_null() || identifier.length == 0) {
+            debug_assert!(
+                identifier.data.is_null(),
+                "unexpected node ID byte string data"
+            );
+            panic!("node ID byte string should have been allocated");
+        }
+
+        Self(inner)
+    }
+
     /// Creates null node ID.
     #[must_use]
     pub(crate) fn null() -> Self {
@@ -71,6 +118,31 @@ impl NodeId {
     #[must_use]
     pub const fn namespace_index(&self) -> u16 {
         self.0.namespaceIndex
+    }
+
+    /// Checks if this node ID is in namespace 0.
+    ///
+    /// Namespace 0 is always the UA namespace `http://opcfoundation.org/UA/` itself and is used for
+    /// fixed definitions as laid out in the OPC UA specification.
+    #[must_use]
+    pub const fn is_ns0(&self) -> bool {
+        self.0.namespaceIndex == 0
+    }
+
+    #[must_use]
+    pub fn is_numeric(&self) -> bool {
+        self.0.identifierType == UA_NodeIdType::UA_NODEIDTYPE_NUMERIC
+    }
+
+    #[must_use]
+    pub fn is_string(&self) -> bool {
+        self.0.identifierType == UA_NodeIdType::UA_NODEIDTYPE_STRING
+    }
+
+    #[must_use]
+    pub fn is_null(&self) -> bool {
+        // SAFETY: Read-only access of static.
+        self == Self::raw_ref(unsafe { &UA_NODEID_NULL })
     }
 
     /// Gets node ID type.